pyproject.toml
[project]
name = "blive-qq-forward"
version = "0.1.0"
description = "BLive -> QQ forwarder"
authors = [{ name = "mokurin000", email = "1348292515a@gmail.com" }]
dependencies = [
    "blivedm@git+https://github.com/mokurin000/blivedm.git@pep621-package",
    "bilibili-api-python>=16.2.0",
    "qq-botpy>=1.2.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


[tool.pdm]
distribution = true
